[inductor][refactor] Extract LazyKernelCompileResult to shared header
Move LazyKernelCompileResult struct from the JIT-only
lazy_triton_compile.h to a shared header at
aoti_include/kernel_compile_result.h. This makes the struct
accessible to both JIT and AOTI compilation paths, which is
needed for the upcoming AOTI lazy Triton compilation support.

Authored with Claude

[ghstack-poisoned]
