+++ torch/csrc/inductor/aoti_include/cuda.h
@@ -1,4 +1,5 @@
 #pragma once
 
 #include <torch/csrc/inductor/aoti_include/common.h>
+#include <torch/csrc/inductor/aoti_include/kernel_compile_result.h>
 #include <torch/csrc/inductor/cpp_wrapper/device_internal/cuda.h>
